+++ python_modules/dagster/dagster_tests/core_tests/test_event_logging.py
@@ -12,7 +12,7 @@ from dagster import (
 
 from dagster.core.events.log import construct_event_logger, EventRecord
 from dagster.core.events import DagsterEventType
-from dagster.core.log import colored_console_logger
+from dagster.core.loggers import colored_console_logger
 
 
 def single_dagster_event(events, event_type):
